add email verifation

+++ app/Http/Controllers/RegisterController.php
@@ -4,14 +4,15 @@ namespace App\Http\Controllers;
 
 
 use App\Models\User;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\http\JsonResponse;
+use Illuminate\Support\Str;
 
 class RegisterController extends BaseController
-{ 
+{
 
     public function register(Request $request)
     {
@@ -29,12 +30,14 @@ class RegisterController extends BaseController
             'name' => $request->name,
             'email' => $request->email,
             'password' => Hash::make($request->password),
+            'mail_token' => Str::random(64),
         ]);
+        // $user->notify(new ConfirmMail($user));
 
-         $token = $user->createToken('frontend-token')->plainTextToken;
+        $token = $user->createToken('frontend-token')->plainTextToken;
 
         return $this->sendResponse([
-             'token' => $token,
+            'token' => $token,
             'user' => $user,
         ], 'Registered successfully');
     }
@@ -47,13 +50,20 @@ class RegisterController extends BaseController
 
         $user = Auth::user();
 
-        
+        if (!$user->email_verified_at) {
+            return $this->sendError(
+                'Please verify your email first',
+                ['email_verified' => false],
+                403
+            );
+        }
+
         $user->tokens()->delete();
 
-         $token = $user->createToken('frontend-token')->plainTextToken;
+        $token = $user->createToken('frontend-token')->plainTextToken;
 
         return $this->sendResponse([
-             'token' => $token,
+            'token' => $token,
             'user' => $user,
         ], 'Login successful');
     }
